Move `TransformCtx` to the `transform` module
