+++ charon/src/transform/index_to_function_calls.rs
@@ -7,7 +7,7 @@ use crate::ids::Vector;
 use crate::llbc_ast::*;
 use crate::meta::Span;
 use crate::pretty::FmtWithCtx;
-use crate::translate_ctx::TransformCtx;
+use crate::transform::TransformCtx;
 use crate::types::*;
 use crate::values::VarId;
 use std::mem::replace;
