+++ charon/src/transform/mod.rs
@@ -1,3 +1,4 @@
+pub mod ctx;
 pub mod graphs;
 pub mod index_to_function_calls;
 pub mod insert_assign_return_unit;
@@ -13,3 +14,5 @@ pub mod reorder_decls;
 pub mod simplify_constants;
 pub mod ullbc_to_llbc;
 pub mod update_closure_signatures;
+
+pub use ctx::TransformCtx;
