+++ charon/src/transform/remove_nops.rs
@@ -4,7 +4,7 @@ use crate::formatter::{Formatter, IntoFormatter};
 use crate::llbc_ast::{RawStatement, Statement};
 use crate::meta::combine_span;
 use crate::pretty::FmtWithCtx;
-use crate::translate_ctx::TransformCtx;
+use crate::transform::TransformCtx;
 use take_mut::take;
 
 fn transform_st(s: &mut Statement) {
