+++ charon/src/transform/remove_unused_locals.rs
@@ -7,7 +7,7 @@ use crate::formatter::{Formatter, IntoFormatter};
 use crate::ids::Vector;
 use crate::llbc_ast::{MutAstVisitor, SharedAstVisitor, Statement};
 use crate::pretty::FmtWithCtx;
-use crate::translate_ctx::TransformCtx;
+use crate::transform::TransformCtx;
 use crate::types::{MutTypeVisitor, SharedTypeVisitor};
 use crate::ullbc_ast::Var;
 use crate::values::*;
