+++ charon/src/transform/reorder_decls.rs
@@ -1,8 +1,8 @@
 use crate::common::*;
 use crate::formatter::{AstFormatter, IntoFormatter};
 use crate::graphs::*;
+use crate::transform::TransformCtx;
 pub use crate::translate_ctx::AnyTransId;
-use crate::translate_ctx::TransformCtx;
 use crate::types::*;
 use crate::ullbc_ast::*;
 use hashlink::linked_hash_map::LinkedHashMap;
